refactor: extract ChartTab from ChartPropertiesPanel
